Exercise8

Documentation for Exercise8. This package implements the zero-dimensional energy balance models (EBMs) introduced in the lecture. The models are coded so that different 0D-EBMs can be easily composed. E.g. via

using Exercise8, OrdinaryDiffEq

model = ZeroDEBM(R_in=IncomingRadiation(α=TanhAlbedo()), R_out=StefanBoltzmannRadiation()) 

prob = ODEProblem(model, 270., (0.,4.), [])
sol = solve(prob, Tsit5())

Exercise8.AbstractEBM

Exercise8.AbstractIncomingRadiation

Exercise8.AbstractOutgoingRadiation

Exercise8.AbstractZeroDEBM

Exercise8.BudykoRadiation

Exercise8.IncomingRadiation

Exercise8.StefanBoltzmannRadiation

Exercise8.TanhAlbedo

Exercise8.ZeroDEBM

Exercise8.AbstractEBM — Type

AbstractEBM{T}

Supertype of all EBMs.

Exercise8.AbstractIncomingRadiation — Type

AbstractIncomingRadiation{T}

Supertype of all incoming radiation / energy schemes.

Exercise8.AbstractOutgoingRadiation — Type

AbstractOutgoingRadiation{T}

Sueprtype of all outgoing radiation / engery schemes.

Exercise8.AbstractZeroDEBM — Type

AbstractZeroDEBM{T} <: AbstractEBM{T}

Supertype of all zero dimensional EBMs.

Exercise8.BudykoRadiation — Type

BudykoRadiation{T} <: AbstractOutgoingRadiation{T}

Implements the outgoing radiation given by Budyko's linear approach

\[\begin{aligned} R_out(T) = A + B\cdot T \end{aligned}\]

By default $A = 202 Wm^{-2}$ and $B = 1.45 W m^{-2} C^{-1}$.

Exercise8.IncomingRadiation — Type

IncomingRadiation{T,F} <: AbstractIncomingRadiation

Implements the incoming radiation given by the formula

\[\begin{aligned} R_in(T,p) = (1 - α(T,p)) \frac{S_0}{4} \end{aligned}\]

Initialized by handing over an albedo, that is a callable function (T,p) -> α or one of the albedo implemented in this package like TanhAlbedo.

Exercise8.StefanBoltzmannRadiation — Type

StefanBoltzmannRadiation{T} <: AbstractOutgoingRadiation{T}

Implements the outgoing radiation given by the Stefan Boltzmann law, potentially reduced by a Greenhouse factor ε

\[\begin{aligned} R_out(T) = \varepsilon\sigma T^4 \end{aligned}\]

Exercise8.TanhAlbedo — Type

TanhAlbedo{T}

Initializes the albedo to depend on temperature with a $tanh$ function

\[\begin{aligned} \alpha(T) = \alpha_0 - \frac{\alpha_1}{2} (\tanh(T - \hat{T})) \end{aligned}\]

Exercise8.ZeroDEBM — Type

ZeroDEBM{T} <: AbstractZeroDEBM

Fields

R_in::AbstractIncomingRadiation

R_out::AbstractOutgoingRadiation